sc-7330 GDS User UI: refact certificate management component
