+++ web/gds-user-ui/src/components/CertificateManagement/TestnetCertificates.tsx
@@ -1,0 +1,5 @@
+function TestnetCertificates() {
+  return <div>TestnetCertificates</div>;
+}
+
+export default TestnetCertificates;
